Add repository layer Move some entity interface implementations to that layer
