+++ middlewares/auth.middleware.js
@@ -1,9 +1,8 @@
 /** @typedef {import("../helpers/types").RequestHandler} RequestHandler */
-const jwt = require("jsonwebtoken");
-const util = require("util");
 const { UnauthorizedError } = require("../helpers/errors");
 const { authentication } = require("../helpers/settings");
 const { filterObj } = require("../helpers/tools");
+const jwt = require("../repositories/jwt.repo");
 
 /**
  * This middleware ensures the user is logged in
@@ -15,9 +14,6 @@ function authGate(req, _, next) {
     .catch((err) => next(new UnauthorizedError(err)));
 }
 
-/** Asyncrhonous version of jwt.verify, based on promises */
-const jwtVerifyAsync = util.promisify(jwt.verify);
-
 /**
  * Asyncrhonous authentication
  * @type {RequestHandler}
@@ -26,7 +22,7 @@ async function authenticate(req) {
   // Extracting the token. Can fail on this stage.
   const token = extractToken(req.get("Authorization"));
   // Verifying token. Can also fail with specific jwt error.
-  const credentials = await jwtVerifyAsync(token, authentication.jwtSecret);
+  const credentials = await jwt.verify(token, authentication.jwtSecret);
   // Searching for user by credentials. Throwing error on fail.
   const user = await req.services.user.getById(credentials.id);
   // The last failsafe — checking if this user actually has the token
